+++ src/Model/Agents/Prey.cpp
@@ -9,7 +9,7 @@ Prey::Prey(const MarkovBrain& brain_,
            uint32_t turn_speed,
            uint32_t segments,
            float los,
-           float fov,
+           uint32_t fov,
            SDL_Texture* sprite_)
     : Agent(brain_, speed, turn_speed, segments, los, fov, sprite_) {}
 Prey::~Prey() {}
